requirements.txt
keras
sklearn
arn.preprocessing
librosa
json
numpy
matplotlib.pyplot 
tensorflow==2.2.0
pandas
seaborn
glob 
os
tqdm
pickle
IPython.display